crud/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
distro==1.8.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.2
openai==1.2.3
pandas==2.1.3
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.1
